Change .exe to .test.exe

+++ Homework-Nov-08/Task2-Counting_rhyme/Task2-Counting_rhyme/CountingRhyme.h
@@ -7,4 +7,4 @@ typedef enum
     countingRhymeExecutedUrgently
 } CountingRhymeErrorCode;
 
-CountingRhymeErrorCode getLastElementValue(const int n, const int m, int* const lastElementValue);
+CountingRhymeErrorCode getLastElementValue(const size_t n, const size_t m, int* const lastElementValue);
